refactor: Use shared logging module
