+++ cmd/unikraft/main.go
@@ -21,7 +21,7 @@ import (
 	"unikraft.com/cli/internal/cmd"
 	"unikraft.com/cli/internal/colors"
 	"unikraft.com/cli/internal/config"
-	"unikraft.com/cli/internal/log"
+	"unikraft.com/x/log"
 )
 
 func main() {
